--- tests/platform/test_warm_reboot.py
@@ -1,7 +1,0 @@
-import pytest
-
-@pytest.mark.usefixtures('get_advanced_reboot')
-def test_warm_reboot_sad(request, get_advanced_reboot):
-
-    advancedReboot = get_advanced_reboot(rebootType='warm-reboot')
-    advancedReboot.runRebootTestcase()
